extract DefaultBubbleHelper class
